pyproject.toml
[tool.poetry]
name = "hoteldjango"
version = "0.1.0"
description = ""
authors = ["ma.minnahmetov <gtr425gtr@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.1"
plotly = "^5.10.0"
pandas = "^1.5.0"
aiohttp = "^3.8.3"
fdb = "^2.0.2"
django-constance = "^2.9.1"
pylint = "^2.15.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
